+++ rundeckapp/grails-spa/src/pages/repository/components/PluginCard.vue
