+++ rundeckapp/grails-spa/src/pages/repository/views/PluginConfigurationView.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>PluginConfigurationView</div>
+</template>
+
+<script>
+import axios from "axios";
+
+export default {
+  name: "PluginConfigurationView"
+};
+</script>
+
+<style lang="scss" scoped>
+</style>
